Reimplement the dashboard to work with DnD and a different model
The logic for working with this is that the Dashboard will be full sized
with slots filled with placeholders over which one can drop widgets.

The slots of the dashboard can be increased and decreased. The slots can't
be decreased if there are widgets that would be out of bounds after that.

The widgets can be resized to span over multiple slots if a valid width,
height is provided.